Update/fix the Windows build
* Retarget Visual C++ platform version to 10.0.19041.0
* Work around unknown attribute error with recent VC2017
* Fix two conversion from double to float loss of data warnings

+++ src/bzflag/Player.cxx
@@ -972,8 +972,8 @@ void Player::addToScene(SceneDatabase* scene, TeamColor effectiveTeam,
     else if (isExploding() && (state.pos[2] > ZERO_TOLERANCE))
     {
         float t;
-        const float deltaExplode = TimeKeeper::getTick() - explodeTime;
-        const float bzExplodeTim = BZDB.eval(StateDatabase::BZDB_EXPLODETIME);
+        const double deltaExplode = TimeKeeper::getTick() - explodeTime;
+        const double bzExplodeTim = BZDB.eval(StateDatabase::BZDB_EXPLODETIME);
         if (deltaExplode < 0)
         {
             // shouldn't happen but why take chances
